pyproject.toml
[project]
name = "marvelrivals-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "goodgd"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django (>=5.1.6,<6.0.0)",
    "pytest (>=8.3.4,<9.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-extensions (>=3.2.3,<4.0.0)",
    "django-debug-toolbar (>=5.0.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
